Add exercise 4.21 solution

+++ part4/controllers/blogs.js
@@ -34,6 +34,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  if (!user) {
+    return response.status(400).json({ error: 'UserId missing or not valid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog && blog.user.toString() !== user._id.toString()) {
+    return response.status(403).json({ error: 'Only the user who added the blog can delete it' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
